gateway working with new consensus sites
